refactor(ai-diagnostic): extract healing_models from self_healing
Moved the action/status/risk enums and the HealingAction / ExecutionResult
records into service/healing_models.py; self_healing.py re-exports them so
importers are unchanged. Decomposed under the existing green suite — the 22
self_healing tests (models + submit/approve/cancel/stats workflow) stay green,
no new tests needed. Dropped now-dead imports (dataclass, field, Enum).
889 -> 799 LOC.
